# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hickory Docs API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/docs/{id}/check": {
      "post": {
        "tags": [
          "runs"
        ],
        "operationId": "check_doc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "doc id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "verification run started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/docs/{id}/run": {
      "post": {
        "tags": [
          "runs"
        ],
        "operationId": "run_doc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "doc id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "optional cell subset (v0 ignores it)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "run started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "the caller's projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/docs": {
      "get": {
        "tags": [
          "projects"
        ],
        "operationId": "list_docs",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "project id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "docs in the project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocSummaryOut"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "operationId": "create_doc",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "project id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "doc created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "operationId": "get_run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "run id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "run status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateDoc": {
        "type": "object",
        "required": [
          "path",
          "source"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        }
      },
      "CreateProject": {
        "type": "object",
        "required": [
          "name",
          "visibility"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          }
        }
      },
...
      "DocOut": {
        "type": "object",
        "description": "`GET /api/docs/:id`, `PUT /api/docs/:id`, and the doc-creation responses\nin `routes::projects` all return this exact shape.",
        "required": [
          "id",
          "path",
          "source",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "DocSummaryOut": {
        "type": "object",
        "required": [
          "id",
          "path",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "ProjectOut": {
        "type": "object",
        "required": [
          "id",
          "name",
          "visibility",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          }
        }
      },
...
      "RunOut": {
        "type": "object",
        "description": "`blocks` is the persisted `[{exec_id, status, transcript}]` projection\n(`crate::runs::run_blocks_from_model`) — kept generic rather than\ndeserialized back into a struct the handler doesn't otherwise need.",
        "required": [
          "id",
          "status",
          "started_at",
          "blocks"
        ],
        "properties": {
          "blocks": {
            "type": "object"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "RunRequest": {
        "type": "object",
        "properties": {
          "cells": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Requested cell subset. v0 executes the whole document regardless:\nexec blocks form a dependency DAG, so partial runs need a dependency\nclosure the pipeline does not expose yet. Accepted for forward\ncompatibility."
          }
        }
      },
      "RunStartOut": {
        "type": "object",
        "required": [
          "run_id"
        ],
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}